changed I2C address and number of bytes request to byte type variable. Updated examples

+++ ADS1110.h
@@ -188,7 +188,7 @@ namespace Ads1110 {
 
     class ADS1110 {
         public:
-            ADS1110(int address);
+            ADS1110(byte address);
             ~ADS1110();
             byte   ping();
             byte   getGain();
